refactor: move DateRange and GroupByPeriod enums to src/common/dto/analytics.dto
