+++ backend/src/common/dto/analytics.dto.ts
@@ -1,0 +1,20 @@
+export enum DateRange {
+  TODAY = 'today',
+  THIS_WEEK = 'this_week',
+  THIS_MONTH = 'this_month',
+  THIS_QUARTER = 'this_quarter',
+  THIS_YEAR = 'this_year',
+  LAST_WEEK = 'last_week',
+  LAST_MONTH = 'last_month',
+  LAST_QUARTER = 'last_quarter',
+  LAST_YEAR = 'last_year',
+  CUSTOM = 'custom',
+}
+
+export enum GroupByPeriod {
+  DAY = 'day',
+  WEEK = 'week',
+  MONTH = 'month',
+  QUARTER = 'quarter',
+  YEAR = 'year',
+}
